JP2OpenJPEG: refactor into re-usable template
Once this PR is validated, it will be re-used in
https://github.com/OSGeo/gdal/pull/8133
